Add Validations and Graphs

+++ FinanceManager/Views/MainMenu.cs
@@ -32,33 +32,35 @@ namespace FinanceManager
                 addFinance.Activate();
                 addFinance.ShowDialog();
                 financeDetails = AddFinance.oFinanceDetails;
-
-                foreach (var item in financeDetails)
+                if (financeDetails != null)
                 {
-                    FinanceManagerDB.FinanceRow row = financeManagerDB.Finance.NewFinanceRow();
-                    row.ContactId = item.ContactId;
-                    row.Name = item.Name;
-                    row.Type = item.Type;
-                    row.Amount = item.Amount;
-                    row.Recurring = item.Recurring;
-                    row.Date = item.Date;
+                    foreach (var item in financeDetails)
+                    {
+                        FinanceManagerDB.FinanceRow row = financeManagerDB.Finance.NewFinanceRow();
+                        row.ContactId = item.ContactId;
+                        row.Name = item.Name;
+                        row.Type = item.Type;
+                        row.Amount = item.Amount;
+                        row.Recurring = item.Recurring;
+                        row.Date = item.Date;
 
-                    financeManagerDB.Finance.AddFinanceRow(row);
-                    financeManagerDB.AcceptChanges();
-                }
+                        financeManagerDB.Finance.AddFinanceRow(row);
+                        financeManagerDB.AcceptChanges();
+                    }
 
-                financeManagerDB.WriteXml("FinanceDatabase.xml");
+                    financeManagerDB.WriteXml("FinanceDatabase.xml");
 
-                FinanceModel financeModel = new FinanceModel();
-                if (await financeModel.SaveFinanceDetails(financeDetails))
-                {
-                    MessageBox.Show("Finance Data Successfully Added!!!");
-                    financeManagerDB = new FinanceManagerDB();
-                    File.Delete("FinanceDatabase.xml");
-                }
-                else
-                {
-                    MessageBox.Show("Something Went Wrong!!!");
+                    FinanceModel financeModel = new FinanceModel();
+                    if (await financeModel.SaveFinanceDetails(financeDetails))
+                    {
+                        MessageBox.Show("Finance Data Successfully Added!!!");
+                        financeManagerDB = new FinanceManagerDB();
+                        File.Delete("FinanceDatabase.xml");
+                    }
+                    else
+                    {
+                        MessageBox.Show("Something Went Wrong!!!");
+                    }
                 }
             }
             catch (Exception ex)
@@ -76,28 +78,31 @@ namespace FinanceManager
                 addContact.ShowDialog();
                 contactDetails = AddContact.oContactDetails;
 
-                foreach (var item in contactDetails)
+                if (contactDetails != null)
                 {
-                    FinanceManagerDB.ContactRow row = financeManagerDB.Contact.NewContactRow();
-                    row.Name = item.Name;
-                    row.Type = item.Type;
-                    row.Category = item.Catogery;
-                    financeManagerDB.Contact.AddContactRow(row);
-                    financeManagerDB.AcceptChanges();
-                }
+                    foreach (var item in contactDetails)
+                    {
+                        FinanceManagerDB.ContactRow row = financeManagerDB.Contact.NewContactRow();
+                        row.Name = item.Name;
+                        row.Type = item.Type;
+                        row.Category = item.Catogery;
+                        financeManagerDB.Contact.AddContactRow(row);
+                        financeManagerDB.AcceptChanges();
+                    }
 
-                financeManagerDB.WriteXml("FinanceDatabase.xml");
+                    financeManagerDB.WriteXml("FinanceDatabase.xml");
 
-                ContactModel contactModel = new ContactModel();
-                if (await contactModel.SaveContactDetails(contactDetails))
-                {
-                    MessageBox.Show("Contact Data Successfully Added!!!");
-                    financeManagerDB = new FinanceManagerDB();
-                    File.Delete("FinanceDatabase.xml");
-                }
-                else
-                {
-                    MessageBox.Show("Something Went Wrong!!!");
+                    ContactModel contactModel = new ContactModel();
+                    if (await contactModel.SaveContactDetails(contactDetails))
+                    {
+                        MessageBox.Show("Contact Data Successfully Added!!!");
+                        financeManagerDB = new FinanceManagerDB();
+                        File.Delete("FinanceDatabase.xml");
+                    }
+                    else
+                    {
+                        MessageBox.Show("Something Went Wrong!!!");
+                    }
                 }
             }
             catch (Exception ex)
@@ -115,36 +120,53 @@ namespace FinanceManager
                 addEvent.ShowDialog();
                 eventDetails = AddEvent.oEventDetails;
 
-                foreach (var item in eventDetails)
+                if (eventDetails != null)
                 {
-                    FinanceManagerDB.EventRow row = financeManagerDB.Event.NewEventRow();
-                    row.Name = item.Name;
-                    row.Description = item.Description;
-                    row.Type = item.Type;
-                    row.Recurring = item.Recurring;
-                    row.Date = item.Date;
-                    financeManagerDB.Event.AddEventRow(row);
-                    financeManagerDB.AcceptChanges();
-                }
+                    foreach (var item in eventDetails)
+                    {
+                        FinanceManagerDB.EventRow row = financeManagerDB.Event.NewEventRow();
+                        row.Name = item.Name;
+                        row.Description = item.Description;
+                        row.Type = item.Type;
+                        row.Recurring = item.Recurring;
+                        row.Date = item.Date;
+                        financeManagerDB.Event.AddEventRow(row);
+                        financeManagerDB.AcceptChanges();
+                    }
 
-                financeManagerDB.WriteXml("FinanceDatabase.xml");
+                    financeManagerDB.WriteXml("FinanceDatabase.xml");
 
-                EventModel eventModel = new EventModel();
-                if (await eventModel.SaveEventDetails(eventDetails))
-                {
-                    MessageBox.Show("Event Data Successfully Added!!!");
-                    financeManagerDB = new FinanceManagerDB();
-                    File.Delete("FinanceDatabase.xml");
-                }
-                else
-                {
-                    MessageBox.Show("Something Went Wrong!!!");
+                    EventModel eventModel = new EventModel();
+                    if (await eventModel.SaveEventDetails(eventDetails))
+                    {
+                        MessageBox.Show("Event Data Successfully Added!!!");
+                        financeManagerDB = new FinanceManagerDB();
+                        File.Delete("FinanceDatabase.xml");
+                    }
+                    else
+                    {
+                        MessageBox.Show("Something Went Wrong!!!");
+                    }
                 }
             }
             catch (Exception ex)
             {
                 MessageBox.Show($"Error: {ex}");
             }
+        }
+
+        private void btnSearchFinance_Click(object sender, EventArgs e)
+        {
+            SearchFinance searchFinance = new SearchFinance();
+            searchFinance.Activate();
+            searchFinance.ShowDialog();
+        }
+
+        private void btnGenerateReport_Click(object sender, EventArgs e)
+        {
+            GenerateReport generateReport = new GenerateReport();
+            generateReport.Activate();
+            generateReport.ShowDialog();
         }
     }
 }
